Move ValidateZone|Region to internal/wire
