+++ pubsublite/admin.go
@@ -35,7 +35,7 @@ type AdminClient struct {
 // See https://cloud.google.com/pubsub/lite/docs/locations for the list of
 // regions and zones where Cloud Pub/Sub Lite is available.
 func NewAdminClient(ctx context.Context, region string, opts ...option.ClientOption) (*AdminClient, error) {
-	if err := validateRegion(region); err != nil {
+	if err := wire.ValidateRegion(region); err != nil {
 		return nil, err
 	}
 	admin, err := wire.NewAdminClient(ctx, region, opts...)
