+++ pubsublite/internal/wire/message_router_test.go
@@ -17,14 +17,9 @@ import (
 	"fmt"
 	"math/rand"
 	"testing"
+
+	"cloud.google.com/go/pubsublite/internal/test"
 )
-
-type fakeSource struct {
-	ret int64
-}
-
-func (f *fakeSource) Int63() int64    { return f.ret }
-func (f *fakeSource) Seed(seed int64) {}
 
 type fakeMsgRouter struct {
 	multiplier     int
@@ -42,7 +37,7 @@ func (f *fakeMsgRouter) Route(orderingKey []byte) int {
 func TestRoundRobinMsgRouter(t *testing.T) {
 	// Using the same msgRouter for each test run ensures that it reinitializes
 	// when the partition count changes.
-	source := &fakeSource{}
+	source := &test.FakeSource{}
 	msgRouter := &roundRobinMsgRouter{rng: rand.New(source)}
 
 	for _, tc := range []struct {
@@ -62,7 +57,7 @@ func TestRoundRobinMsgRouter(t *testing.T) {
 		},
 	} {
 		t.Run(fmt.Sprintf("partitionCount=%d", tc.partitionCount), func(t *testing.T) {
-			source.ret = tc.source
+			source.Ret = tc.source
 			msgRouter.SetPartitionCount(tc.partitionCount)
 			for i, want := range tc.want {
 				got := msgRouter.Route([]byte("IGNORED"))
